fix(playback-rate-button): update rates attr parsing
Previously, rates was parsing as `1, 2, 3` but it was supposed to be
parsing as an HTML-style list like `1 2 3`. This updates the regex for
the split from `,\s?` to `\s+|,\s*` to support what is documented but
also keep around the legacy behavior.

In addition, the map was changed to use `Number(str)` for converting the
strings into numbers and the sorting was made to be numerical.