fix: use mouseleave instead of mouseout
With mouseout, we would get an event each time we cross an element
boundary, meaning we could end up marking the player as inactive as a
user is mousing around in the player. Instead, we should use mouseleave
that would only trigger when we leave the entire element including its
children.