fix: make media-time-display not wrap when small
media-time-display shows both the current or remaining time and the
duration separated by a slash. Around the slash, there are spaces, so,
when the control bar is small, this causes the element to wrap onto
multiple lines. This can cause unexpected styling issues with the
control bar.

The expectation of this element is to stay on one-line like the other
time displays.

While we could add this into media-text-display, it's already used for
non time displays in the examples, though, internally in media-chrome,
media-text-display is only used in the time displays.