feat: re-enable media click to play toggle
This enables clicking on the media element to toggle playback. This
works on both desktop and mobile/touch devices. Eventually, we'd probably want
to make mobile/touch devices to toggle controls showing on taps (except
for the first one).

One thing that's worth noting, in the mobile example, if during
playback, you click the play toggle, the controls will stay showing.
This is actually an existing issue relating to us keeping controls
showing when focus is on a control bar control (like the
`media-play-button`).