fix: unhide controls on touch & autohide on timer
This keeps desktop behavior for mobile/touch devices. When
tapping, it'll unhide the controls and then setup the timer to autohide.

`mousemove` triggers even when a user only does a single tap.
`pointermove` doesn't occur with just a single tap, so, we switch
`mousemove` to `pointermove`. Then, we schedule the inactivity timeout
on `pointerup` when `pointerType` is `touch`, since other pointer types
are covered.