fix: hide gesture-layer when in audio mode
The gestures layer, depending on configuration, may add an extra 5
pixels above or below the control bar, which is unwanted. Instead, when
we're playing back audio, we should hide the gestures display, which
involves making it a block element so that we can set the height of it
to zero.