feat: keyboard shortcut support (#263)
When the focus is on the media-controller or a child element of it, pressing f,k,m,Space,ArrowLeft,ArrowRight will toggle fullscreen, toggle playback, toggle mute, toggle playback, seek back 10s, seek forward 10s respectively.

If an interactive element is already bound to one of those keys, its behavior would be prioritized and the keyboard shortcut's behavior wouldn't trigger. If implementing manually, expose a `keysUsed` property or add a `keysused` attribute to the element.

In addition, to turn off keyboard shortcuts behavior, add the `nohotkeys` attribute to the media-controller.