fix: prevent hotkeys from scrolling page
If a key will scroll the page, such as Space, ArrowLeft, and ArrowRight,
we want to prevent the default action so the page won't scroll.

Browsers scroll on keydown rather than keyup, so, we have to do the
preventDefault separately from where we do the action. This has the
side-effect that should someone choose to cancel an action --
for example, holding down space and then hitting escape to prevent
playback toggling from happening -- then the default behavior, in the
example, scrolling down a page, will still be prevented.

Fixes #295