fix: set tabindex on media element, if not set
This is so that clicking inside the player gets focus changed to be
inside the player and hotkeys work as expected. If a tabindex is already
set, we shouldn't override it. Otherwise, we should set it to be -1 so
that clicking would change focus but it wouldn't be added to the regular
tab order.

Since the primary usecase of this is for hotkeys, there will be keyboard
interactions with the player after the focus change. However, this does
mean that an outline will be added to the media element. This probably
isn't an issue when windowed, but in fullscreen, we definitely don't
want an outline border on the media.

Fixes #309