feat: introduce a playback rates selectmenu (#513)
This uses a playback-rate-listbox under the hood which accepts items declaratively only.
```html
<media-playback-rate-listbox media-controller="mc">
  <media-chrome-listitem value="1">1x</media-chrome-listitem>
  <media-chrome-listitem value="1.5">1.5x</media-chrome-listitem>
  <media-chrome-listitem value="2">2x</media-chrome-listitem>
  <media-chrome-listitem value="3">3x</media-chrome-listitem>
</media-playback-rate-listbox>
```
The selectmenu itself accepts a rates property like the playback-rate-button does. It also defaults to the same defaults.
```html
<media-playback-rate-selectmenu rates="1 2 3 4"></media-playback-rate-selectmenu>
```
This also has some updates to the main listbox class such has the value getter/setter, which is available on the HTMLSelectElement as a both as well.

In addition, it changes the behavior of the listitem toggling to only toggle with multi-select or with command/ctrl click.