feat: Icon slots for buttons with multiple slots (#643)
Adds an optional `icon` slot to buttons with multiple available slots
e.g. Captions, Cast, Fullscreen, PiP and Play.

Updates docs with basic examples of usage. There's also a new
`animated-icons.html` in the examples folder but this also currently has
very basic examples to allow us to check the slots work as expected. We
can expand this in the future with fancier animated demos.