fix: time range preview and value inconsistency (#670)
Fixes https://github.com/muxinc/media-chrome/issues/656

As it turned out, input[type=range] makes half the width of thumb a
"dead zone" where the value is either max or min. If we account for that
when calculating preview values, we get the "right" result in preview.
On a video of 3 minutes, the difference is < 0.01s between preview and
input value.