fix: media-chrome-option outline fix and new css vars (#722)
Adds new CSS vars for controlling hover and non hover outlines. This
fixes an issue where outlines are being cut off on the left and right
side of the options.

Hover outline styles are also applied to the `:focus-visible` state for
consistency.

There's also a new `--media-option-focus-shadow` var which is helpful
for disabling the shadow on `:focus-visible` for simply allowing the
option to look the same on hover and focus relying only on the outline
vars.