fix: access sheet before connected to DOM (#813)
same than the previous bug fix. 

good it uncovered some more issues that are related to trying to access
a style sheet before connected to the DOM which seems not possible.

we still keep the dummy CSSStyleRule in the return to not have type
errors but I added a warning and a comment to indicate there might be an
issue in the code.