tools: upgrade esbuild, wet-run serve, livereload (#827)
this PR uses wet-run to serve the vanilla examples demo, makes it so
`yarn dev` watches all build formats (esm, cjs, iife) so those get
updated while we run `yarn dev` in the docs. The IIFE bundle is used in
the Sandpacks.

this also enables `livereload` for the vanilla examples which speeds up
developing.