feat: React MediaStore (#851)
NOTE: Per a conversation with @heff and prompted by @mmcc's gut
example/pseudo-code on importing, at least at this stage, we decided to
call an audible and not create multiple npm packages / restructure our
codebase to a monorepo. As this was decided between just the two of us,
we can certainly discuss more. Additionally, we can merge this as is but
hold off on any "official support" and corresponding comms/documentation
unless/until we have sufficient consensus on single package vs.
multi-package for the new react-specific code, especially since this
should have no impact on prior use cases of Media Chrome.

Demo usage/smoke testing:
https://media-chrome-demo-nextjs-git-fork-cjpillsbury-experi-b8bc5e-mux.vercel.app/