fix: seeking when preload is none (#997)
related #910

this fixes an issue where when you seek before the video is loaded we
don't set the video.currentTime.

it's fine to set the currentTime anytime. it doesn't throw for a valid
number value.
but users have to set the `defaultduration` in some cases. not sure with
iOS and native HLS.