fix: revert theme render flicker (breakpoints) (#1055)
related #960

from feedback via next-video we heard there is a split second empty
flicker now when using a slotted poster image.
first the blurhash appears, after the thumbnail then nothing (seeing
background of page), then the poster and player UI.

the reason is that the theme visibility is hidden until the breakpoints
are available.
after reconsideration this solution doesn't feel like the correct way to
handle this.

better is to be more precise in the theme itself and require breakpoints
to render any control variants.
this will require some changes in some themes where we don't default to
the smallest size variant for example but still require a small
breakpoint width. could be even zero.