fix: remove media container observers on disconnect (#1061)
As discussed in https://github.com/muxinc/media-chrome/pull/1057.

Apologies, not sure why it's auto-formatting the innerHTML template
string, am I missing an extension or something?

To answer a few questions from the other PR:

> valid points though the constructor is called on an upgrade and most
of the time that is on connected.

Generally true, but not when the element is created in-memory
(`document.createElement` or via template) and not yet attached to the
DOM. Quite common practice in modern frameworks (e.g.,
Vue/Solid/Svelte).

> I believe we put it there because it is just setup once but it could
be an oversight. it doesn't cause any serious issues but a new PR
putting it in connectedCallback seems good to me. last I checked the
cleanup logic is not absolutely required because the browser cleans that
up automatically when the element is garbage collected.

That is true but it's still best practice to clean up on disconnect.
Element may be held in memory as it's being moved in the DOM, might be
offloaded in a scroll container as it goes out of view, detached by a
framework as a performance optimization, and many other reasons.