fix: firefox pip activates when clicked through on settings menu (#1074)
In order to fix to this [issue
](https://github.com/muxinc/media-chrome/issues/1069) where in firefox
they have a floating toggle for Pip that interrupt the use of media
chrome menus.

I made a research and we discover that[ Firefox uses their own PiP
system](https://support.mozilla.org/en-US/kb/about-picture-picture-firefox)
adding a toggle overlaying videos and other elements if they dont have
background color to 100% and opacity to 1 this make our menus
unclickeable if they are under this toggle.

For now this can be partially solved adding a css class for background
solid color that only applies into firefox and we created this
[ticket](https://bugzilla.mozilla.org/show_bug.cgi?id=1947825) for get a
better solution from mozilla them since I also discover Youtube had the
same problem but they created an exception that prevent this UX bug for
elements with transparencies from 0.7 or higher for them identifying its
youtube by the domain.

![Screenshot 2025-02-12 at 4 26
29 PM](https://github.com/user-attachments/assets/8d13fb64-dbd2-4f90-a61d-6e6ef1796a6d)