feat: persist muted enabled state (#1088)
Closes #1072
## Changes
- `MEDIA_UNMUTE_REQUEST`: When unmuting, if the preferred volume is too
low (<= 0.25), set the volume to 0.25. Otherwise, restore the preferred
volume.
- `mediaMuted` set behavior: When muting, set `media-chrome-pref-muted`
to `true` in localStorage and when unmuting set it to `false`.
- `stateOwnersUpdateHandlers` behavior: calls `mediaMuted`'s set with
the `mutedPref` value.

I replicated the same user experience as yt when muting and unmuting
(restoring the previously selected volume), with the exception that if
the preferred volume is too low (<=0.25), it is set to 0.25 to ensure
the audio is clearly perceptible. I'm open to any other suggestions for
doing this :)

---------

Co-authored-by: Wesley Luyten <me@wesleyluyten.com>