feat: hide-cursor-controls-on-inactivity (#1101)
This PR resolves [issue
#756](https://github.com/muxinc/media-chrome/issues/756), addressing the
case where controls and the cursor should hide even when hovering over
the controls.

### Changes Introduced
- Added the alwaysHideControls attribute to the media controller.
- When this attribute is present, controls and the cursor will hide
after user inactivity, even if the user is hovering over the controls
only if video is playing and not casting.

#### Behavior
##### With alwaysHideControls:
- After clicking play, if the user does not move the cursor, both the
controls and cursor will hide—even when hovering over the controls.
#### Without alwaysHideControls (default behavior):
- Controls and cursor will only hide if the user is not hovering over
the controls is playing and is not casting