feat: set i18n through media-controller (#1105)
This adds ~~a singleton class~~ `let` to manage the language, and uses
the browser language as fallback. If the browser language isn't present
as language file, it'll just grab the `en`, like it did before.

The language for this can be set through an attribute on
media-controller like in the example. So:
```html
<media-controller language="fr"></media-controller>
```

Now this can be managed outside media-chrome through a cookie, interface
element or whatsoever.

---------

Co-authored-by: Wesley Luyten <me@wesleyluyten.com>