Hide fullscreen button for <cloudflare-video> on iOS due to platform limitations (#1129)
This PR addresses
[#1124](https://github.com/muxinc/media-chrome/issues/1124).

On iOS browsers (Safari, Brave, Chrome), the fullscreen button for
<cloudflare-video> does not function properly because of platform
restrictions on fullscreening iframe-based elements.

To improve the user experience and avoid presenting a non-functional
control, the fullscreen button is now hidden when fullscreen is unable
iOS devices.

Also add console error when tries to fullscreen and it's not posible.