docs: clarify <media-preview-thumbnail> needs a VTT track and add reference link (#1187)
Resolves
[#1179](https://github.com/muxinc/media-chrome/issues/1179#issuecomment-3204189445)
where a user reported confusion about why the component wasn’t
displaying thumbnails by default. This update makes the requirement for
a VTT track explicit and provides a direct link to the samples in
[media-time-range
](https://www.media-chrome.org/docs/en/components/media-time-range#preview-thumbnails)for
further details.

This PR updates the documentation for `<media-preview-thumbnail>` to
clarify that preview thumbnails only appear when a metadata text track
labeled “thumbnails” is provided with a valid VTT file.